Movin code outside the service
