+++ src/Models/Product/ProductService.ts
@@ -1,12 +1,19 @@
 import * as mongoose from "mongoose";
+import AbstractEntityService from "../../Base/AbstractEntityService";
 import ProductSchema from "./ProductSchema";
 
-export default class ProductService implements EntityService {
+export default class ProductService extends AbstractEntityService implements EntityService {
 
     protected productSchema: any;
 
     constructor() {
+        super();
+
         this.productSchema = mongoose.model("Product", ProductSchema);
+    }
+
+    public getSchema() {
+        return this.getProduct();
     }
 
     public setProduct(product: any) {
@@ -15,36 +22,6 @@ export default class ProductService implements EntityService {
 
     public getProduct() {
         return this.productSchema;
-    }
-
-    public async getAll() {
-        return await this.productSchema.find({}).lean();
-    }
-
-    public async load(id: string) {
-        return this.productSchema.findById(id).lean();
-    }
-
-    public update(id: string, values: any, callback: any = null) {
-        return this.productSchema.findOneAndUpdate({_id: id}, values, {new: true}, callback).exec();
-    }
-
-    public async create(object: object) {
-        const product = new this.productSchema(object);
-
-        return await product.save();
-    }
-
-    public async delete(id: string) {
-        const loadedProduct = await this.productSchema.findById(id);
-
-        if (!loadedProduct) {
-            return new Promise((resolve, reject) => {
-                reject("item_not_exists");
-            });
-        }
-
-        return await this.productSchema.deleteOne({_id: id});
     }
 
 }
